This is The Odin Project Assignment.
This commit is for the HTML of the assignment.

In this assignment I have been dealing with Headings, paragraphs,
ordered and unordered lists, links and images.

At this stage am glad for completing this challange I would and so eager
to conrinue learning, hopefully i will understand also the CSS and
Javascript concepts.

Understanding how to work with git and github.

More understanding on using VScode as a text editor.

Also I got more familiar with HTML code.